+++ webpack.config.js
@@ -3,16 +3,16 @@ const path = require('path');
 const webpack = require('webpack');
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 
-function createBanner(){
+function createBanner(withTzdbVersion = true){
     const packageJson = require('./package.json');
     const tzdbLatest = require('moment-timezone/data/packed/latest');
 
-    const version = '//! @version ' + packageJson.name + ' - ' + packageJson.version + '-' + tzdbLatest.version + '\n';
+    const version = withTzdbVersion ?
+        `//! @version ${packageJson.name} - ${packageJson.version} - ${tzdbLatest.version}\n` :
+        `//! @version ${packageJson.name} - ${packageJson.version}\n`;
     const preamble = fs.readFileSync('./src/license-preamble.js', 'utf8');
     return version + preamble;
 }
-
-const banner = createBanner();
 
 const externals = {
     'js-joda': {
@@ -48,9 +48,9 @@ const optimization = {
     ]
 };
 
-const plugins = [
+const plugins = withTzdbVersion => [
     new webpack.BannerPlugin(
-        {banner: banner, raw: true}
+        {banner: createBanner(withTzdbVersion), raw: true}
     ),
 ];
 
@@ -68,7 +68,24 @@ const productionConfig = {
     externals,
     module: moduleConfig,
     optimization,
-    plugins,
+    plugins: plugins(true),
+};
+
+const productionConfigEmpty = {
+    mode: 'production',
+    context: __dirname,
+    entry: './src/js-joda-timezone-empty.js',
+    devtool: false,
+    output: {
+        path: __dirname  + '/dist',
+        filename: 'js-joda-timezone-empty.min.js',
+        libraryTarget: 'umd',
+        library: 'JSJodaTimezone'
+    },
+    externals,
+    module: moduleConfig,
+    optimization,
+    plugins: plugins(false),
 };
 
 const developmentConfig = {
@@ -84,9 +101,30 @@ const developmentConfig = {
     },
     externals,
     module: moduleConfig,
-    plugins,
+    plugins: plugins(true),
 };
 
-const config = [developmentConfig, productionConfig];
+const developmentConfigEmpty = {
+    mode: 'development',
+    context: __dirname,
+    entry: './src/js-joda-timezone-empty.js',
+    devtool: 'hidden-source-map',
+    output: {
+        path: __dirname  + '/dist',
+        filename: 'js-joda-timezone-empty.js',
+        libraryTarget: 'umd',
+        library: 'JSJodaTimezone'
+    },
+    externals,
+    module: moduleConfig,
+    plugins: plugins(false),
+};
+
+const config = [
+    developmentConfig,
+    productionConfig,
+    developmentConfigEmpty,
+    productionConfigEmpty,
+];
 
 module.exports = config;